+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-* Added: ForiNet client VPN cmdlets
+* Added: FortiNet client VPN cmdlets
 
 
 ## 2.4.0 - 2018-10-01